Reorganize, add some tests
