+++ frontend/utilities/helpers.ts
@@ -899,31 +899,3 @@ export default {
   normalizeEmptyValues,
   wrapFleetHelper,
 };
-
-export const STYLIZATIONS_AND_ACRONYMS = [
-  "macOS",
-  "osquery",
-  "MySQL",
-  "MDM",
-  "REST",
-  "API",
-  "JSON",
-];
-
-// fleetdm.com/handbook/marketing/content-style-guide#sentence-case
-// * doesn't recognize proper nouns!
-export const enforceFleetSentenceCasing = (s: string) => {
-  const resArr = s.split(" ").map((word, i) => {
-    if (!STYLIZATIONS_AND_ACRONYMS.includes(word)) {
-      const lowered = word.toLowerCase();
-      if (i === 0) {
-        // title case the first word
-        return lowered[0].toUpperCase() + lowered.slice(1);
-      }
-      return lowered;
-    }
-    return word;
-  });
-
-  return resArr.join(" ");
-};
